test(chore): only remove workspace if it exists

+++ test/test_helper.bash
@@ -10,7 +10,7 @@ function setup {
 # reset the test state
 function clean_test {
 	cd $TEST_ROOT
-	rm -R workspace/
+	[[ -d workspace/ ]] && rm -R workspace/
 	mkdir workspace
 	chmod 777 workspace/
 	mkdir workspace/cwd
